fix: bound CFF narrowing casts (V-038/V-045/V-049)
Three correctness LOWs in the CFF parser where a silent narrowing cast
turned attacker-controlled out-of-range input into a valid-looking
in-range value:

- V-038 (CFFFileInput::Type2Endchar): the deprecated seac flavor of
  endchar takes bchar/achar character codes in [0, 255]. The operands
  were cast to Byte unconditionally, so 0x1234 silently became 0x34 and
  dependency calculation walked into the wrong StandardEncoding glyph.
  Cast operands to long first, bound-check [0, 255], return eFailure +
  TRACE_LOG on out-of-range. Behavior unchanged for legitimate fonts.

- V-045 (CFFFileInput::GetBiasedIndex): a malicious subr index could
  wrap through the unsigned-short cast into a different valid subr
  (subset substitution). Changed the (private) helper to return long
  so callers can range-check both ends; GetLocalSubr / GetGlobalSubr
  now reject negative and out-of-range biased indices explicitly
  instead of relying on a one-sided unsigned compare.

- V-049 (CFFPrimitiveReader::Read4ByteSigned): `(int)buffer` assumed
  int is exactly 32 bits to get the sign right. Use `(int32_t)` from
  <stdint.h> so 32-bit sign extension into long is guaranteed
  regardless of platform.

Test: RunType2EndcharBadOperandCases — six rows covering above-255 /
negative / real-valued bchar and achar operands; all return eFailure
post-fix. V-045 / V-049 have no observable behavior change on
legitimate input, so no new tests; existing CFF suite verifies no
regression.

+++ PDFWriter/CFFFileInput.h
@@ -304,7 +304,11 @@ private:
 												 long inDefault);
 	LongFilePositionType GetCharsetPosition(unsigned short inFontIndex);
 	LongFilePositionType GetEncodingPosition(unsigned short inFontIndex);
-	unsigned short GetBiasedIndex(unsigned short inSubroutineCollectionSize, long inSubroutineIndex);
+	// Returns the biased subroutine index as a signed long so callers can
+	// range-check before deref. Returning unsigned short would silently
+	// wrap attacker-controlled inSubroutineIndex values into a valid
+	// in-range subr slot (subset substitution).
+	long GetBiasedIndex(unsigned short inSubroutineCollectionSize, long inSubroutineIndex);
 	PDFHummus::EStatusCode ReadFormat0Charset(bool inIsCID, UShortToCharStringMap& ioCharMap,unsigned short** inSIDArray,const CharStrings& inCharStrings);
 	PDFHummus::EStatusCode ReadFormat1Charset(bool inIsCID,UShortToCharStringMap& ioCharMap,unsigned short** inSIDArray,const CharStrings& inCharStrings);
 	PDFHummus::EStatusCode ReadFormat2Charset(bool inIsCID,UShortToCharStringMap& ioCharMap,unsigned short** inSIDArray,const CharStrings& inCharStrings);
